fix for events on 1.4.2
The problem was that we do comparsion of the event emitted
contract-package-hash and the one taken from contract named keys.
The first one is all-lowercase as its converted by JS-SDK which doesn't
support checksumed keys and the second one is provided as a string so it
can have mixed case value.
Now we call toLowerCase() on contract-package-hash when client starts so
it matches values provided while calling `setContractHash`